add new file

+++ src/Pages/Cart/Cart.js
@@ -1,5 +1,6 @@
 import React, { createContext, useContext, useEffect, useState } from 'react';
 import { useParams } from 'react-router';
+import Swal from 'sweetalert2';
 import UseFirebase from '../../Firebase/UseFirebase';
 import { productsContext } from '../Home/Products/Products';
 
@@ -16,49 +17,58 @@ const Cart = () => {
 
 
     useEffect(()=>{
-        fetch(`http://localhost:5000/carts/${email}`)
+        fetch(`https://mighty-plains-84607.herokuapp.com/carts/${email}`)
           .then(res => res.json())
           .then(data => {
             setCartTotal(data)
             setProducts(data)
            
           } )
-    }, [products] )
+    }, [products])
 
 
     const handaleDelete = id => {
 
       
-      const uri = `http://localhost:5000/carts/${id}`;
-      console.log(uri);
-      fetch( uri ,{
-            method: "DELETE"
-        })
-        .then()
-        .then(data =>{
-            if(data.deleteCount > 0){
-                const remainingProduct = products.filter(product => product._id !== id)
-                setProducts(remainingProduct);
-            }
-        })
+      const uri = `https://mighty-plains-84607.herokuapp.com/carts/${id}`;
+      const exist = window.confirm("Are You sure want to delete ??")
+
+      if(exist){
+        fetch( uri ,{
+          method: "DELETE"
+      })
+      .then()
+      .then(data =>{
+      
+
+       if(data.deleteCount > 0) {
+            alert("Delete Successfuly");
+              const remainingProduct = products.filter(product => product._id !== id)
+              setProducts(remainingProduct);
+          }
+          
+          
+
+      })
+      }
     }
 
     return (
         <section className="container mt-5 pt-5">
 
-<article className="cart-container gap-4 row ">
+<article className="cart-container gap-4 row">
     
 {
-                products.map(product => <article class="card mb-3" style={{maxWidth: "420px"}}>
+                products.map(product => <article className="card mb-3 shadow" style={{maxWidth: "420px"}}>
               
-                <div class="row g-0">
-                  <div class="col-md-4">
-                    <img src={product?.image} class="img-fluid rounded-start" alt="..." />
+                <div className="row g-0">
+                  <div className="col-md-4">
+                    <img src={product?.image} className="img-fluid rounded-start" alt="..." />
                   </div>
-                  <div class="col-md-8">
-                    <div class="card-body">
-                      <h5 class="card-title">{product?.title}</h5>
-                      <h6 class="card-text text-muted my-3 mb-3">Price: ${product?.price}</h6>
+                  <div className="col-md-8">
+                    <div className="card-body">
+                      <h5 className="card-title">{product?.title}</h5>
+                      <h6 className="card-text text-muted my-3 mb-3">Price: ${product?.price}</h6>
                     <button className="btn btn-warning" onClick={()=> handaleDelete(product?._id)}>REMOVE</button>
                     </div>
                   </div>
